+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -188,7 +188,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
                 // Wait for the child process to upload all the data.
                 executeChild()
                 waitForExpectations(timeout: 10.0, handler: nil)
-                token!.stop()
+                token!.invalidate()
                 XCTAssert(callCount > 1)
                 XCTAssert(transferred >= transferrable)
             } else {
@@ -234,7 +234,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
                 }
             }
             waitForExpectations(timeout: 10.0, handler: nil)
-            token!.stop()
+            token!.invalidate()
             XCTAssert(callCount > 1)
             XCTAssert(transferred >= transferrable)
         } catch {
@@ -335,7 +335,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             let exp = expectation(description: "A new permission offer will be processed by the server")
 
             let results = managementRealm.objects(SyncPermissionOffer.self).filter("id = %@", permissionOffer.id)
-            let notificationToken = results.addNotificationBlock { (changes) in
+            let notificationToken = results.observe { (changes) in
                 if case .update(let change, _, _, _) = changes, let statusCode = change[0].statusCode.value {
                     XCTAssertEqual(statusCode, 0)
                     XCTAssertEqual(change[0].status, .success)
@@ -349,7 +349,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             }
 
             waitForExpectations(timeout: 2)
-            notificationToken.stop()
+            notificationToken.invalidate()
         } catch {
             XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
         }
@@ -370,7 +370,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             let permissionOfferNotificationToken = managementRealm
                 .objects(SyncPermissionOffer.self)
                 .filter("id = %@", permissionOffer.id)
-                .addNotificationBlock { (changes) in
+                .observe { (changes) in
                 if case .update(let change, _, _, _) = changes, let statusCode = change[0].statusCode.value {
                     XCTAssertEqual(statusCode, 0)
                     XCTAssertEqual(change[0].status, .success)
@@ -386,7 +386,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             }
 
             waitForExpectations(timeout: 2)
-            permissionOfferNotificationToken.stop()
+            permissionOfferNotificationToken.invalidate()
 
             let userB = try synchronouslyLogInUser(for: basicCredentials(register: isParent, usernameSuffix: "_B"), server: authURL)
             _ = try synchronouslyOpenRealm(url: realmURL, user: userB)
@@ -403,7 +403,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             let permissionOfferResponseNotificationToken = managementRealm
                 .objects(SyncPermissionOfferResponse.self)
                 .filter("id = %@", permissionOfferResponse.id)
-                .addNotificationBlock { (changes) in
+                .observe { (changes) in
                 if case .update(let change, _, _, _) = changes, let statusCode = change[0].statusCode.value {
                     XCTAssertEqual(statusCode, 0)
                     XCTAssertEqual(change[0].status, .success)
@@ -421,7 +421,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             }
 
             waitForExpectations(timeout: 2)
-            permissionOfferResponseNotificationToken.stop()
+            permissionOfferResponseNotificationToken.invalidate()
 
             _ = try synchronouslyOpenRealm(url: URL(string: responseRealmUrl!)!, user: userB)
         } catch {