+++ Realm/RLMRealm.mm
@@ -67,7 +67,7 @@ void RLMDisableSyncToDisk() {
 }
 
 @implementation RLMRealmNotificationToken
-- (void)stop {
+- (void)invalidate {
     [_realm verifyThread];
     [_realm.notificationHandlers removeObject:self];
     _realm = nil;
@@ -93,7 +93,7 @@ void RLMDisableSyncToDisk() {
     if (_realm || _block) {
         NSLog(@"RLMNotificationToken released without unregistering a notification. You must hold "
               @"on to the RLMNotificationToken returned from addNotificationBlock and call "
-              @"-[RLMNotificationToken stop] when you no longer wish to receive RLMRealm notifications.");
+              @"-[RLMNotificationToken invalidate] when you no longer wish to receive RLMRealm notifications.");
     }
 }
 @end