+++ RealmSwift/Object.swift
@@ -217,7 +217,7 @@ open class Object: RLMObjectBase, ThreadConfined {
 
      Only objects which are managed by a Realm can be observed in this way. You
      must retain the returned token for as long as you want updates to be sent
-     to the block. To stop receiving updates, call `stop()` on the token.
+     to the block. To stop receiving updates, call `invalidate()` on the token.
 
      It is safe to capture a strong reference to the observed object within the
      callback block. There is no retain cycle due to that the callback is
@@ -229,7 +229,7 @@ open class Object: RLMObjectBase, ThreadConfined {
      - parameter block: The block to call with information about changes to the object.
      - returns: A token which must be held for as long as you want updates to be delivered.
      */
-    public func addNotificationBlock(_ block: @escaping (ObjectChange) -> Void) -> NotificationToken {
+    public func observe(_ block: @escaping (ObjectChange) -> Void) -> NotificationToken {
         return RLMObjectAddNotificationBlock(self, { names, oldValues, newValues, error in
             if let error = error {
                 block(.error(error as NSError))