+++ RealmSwift/RealmCollection.swift
@@ -50,7 +50,7 @@ public final class RLMIterator<T: Object>: IteratorProtocol {
  For example, for a simple one-section table view, you can do the following:
 
  ```swift
- self.notificationToken = results.addNotificationBlock { changes in
+ self.notificationToken = results.observe { changes in
      switch changes {
      case .initial:
          // Results are now populated and can be accessed without blocking the UI
@@ -327,7 +327,7 @@ public protocol RealmCollection: RealmCollectionBase {
      ```swift
      let results = realm.objects(Dog.self)
      print("dogs.count: \(dogs?.count)") // => 0
-     let token = dogs.addNotificationBlock { changes in
+     let token = dogs.observe { changes in
      switch changes {
          case .initial(let dogs):
              // Will print "dogs.count: 1"
@@ -349,17 +349,17 @@ public protocol RealmCollection: RealmCollectionBase {
      ```
 
      You must retain the returned token for as long as you want updates to be sent to the block. To stop receiving
-     updates, call `stop()` on the token.
+     updates, call `invalidate()` on the token.
 
      - warning: This method cannot be called during a write transaction, or when the containing Realm is read-only.
 
      - parameter block: The block to be called whenever a change occurs.
      - returns: A token which must be held for as long as you want updates to be delivered.
      */
-    func addNotificationBlock(_ block: @escaping (RealmCollectionChange<Self>) -> Void) -> NotificationToken
+    func observe(_ block: @escaping (RealmCollectionChange<Self>) -> Void) -> NotificationToken
 
     /// :nodoc:
-    func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) -> NotificationToken
+    func _observe(_ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) -> NotificationToken
 }
 
 private class _AnyRealmCollectionBase<T: Object>: AssistedObjectiveCBridgeable {
@@ -389,7 +389,7 @@ private class _AnyRealmCollectionBase<T: Object>: AssistedObjectiveCBridgeable {
     func value(forKey key: String) -> Any? { fatalError() }
     func value(forKeyPath keyPath: String) -> Any? { fatalError() }
     func setValue(_ value: Any?, forKey key: String) { fatalError() }
-    func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
+    func _observe(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { fatalError() }
     class func bridging(from objectiveCValue: Any, with metadata: Any?) -> Self { fatalError() }
     var bridged: (objectiveCValue: Any, metadata: Any?) { fatalError() }
@@ -498,8 +498,8 @@ private final class _AnyRealmCollection<C: RealmCollection>: _AnyRealmCollection
     // MARK: Notifications
 
     /// :nodoc:
-    override func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
-        -> NotificationToken { return base._addNotificationBlock(block) }
+    override func _observe(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
+        -> NotificationToken { return base._observe(block) }
 
     // MARK: AssistedObjectiveCBridgeable
 
@@ -763,7 +763,7 @@ public final class AnyRealmCollection<T: Object>: RealmCollection {
      ```swift
      let results = realm.objects(Dog.self)
      print("dogs.count: \(dogs?.count)") // => 0
-     let token = dogs.addNotificationBlock { changes in
+     let token = dogs.observe { changes in
          switch changes {
          case .initial(let dogs):
              // Will print "dogs.count: 1"
@@ -785,19 +785,19 @@ public final class AnyRealmCollection<T: Object>: RealmCollection {
      ```
 
      You must retain the returned token for as long as you want updates to be sent to the block. To stop receiving
-     updates, call `stop()` on the token.
+     updates, call `invalidate()` on the token.
 
      - warning: This method cannot be called during a write transaction, or when the containing Realm is read-only.
 
      - parameter block: The block to be called whenever a change occurs.
      - returns: A token which must be held for as long as you want updates to be delivered.
      */
-    public func addNotificationBlock(_ block: @escaping (RealmCollectionChange<AnyRealmCollection>) -> Void)
-        -> NotificationToken { return base._addNotificationBlock(block) }
+    public func observe(_ block: @escaping (RealmCollectionChange<AnyRealmCollection>) -> Void)
+        -> NotificationToken { return base._observe(block) }
 
     /// :nodoc:
-    public func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<AnyRealmCollection>) -> Void)
-        -> NotificationToken { return base._addNotificationBlock(block) }
+    public func _observe(_ block: @escaping (RealmCollectionChange<AnyRealmCollection>) -> Void)
+        -> NotificationToken { return base._observe(block) }
 }
 
 // MARK: AssistedObjectiveCBridgeable
@@ -823,7 +823,12 @@ extension AnyRealmCollection: AssistedObjectiveCBridgeable {
 
 // MARK: Unavailable
 
-extension AnyRealmCollection {
+extension RealmCollection {
     @available(*, unavailable, renamed: "sorted(byKeyPath:ascending:)")
     func sorted(byProperty property: String, ascending: Bool) -> Results<Element> { fatalError() }
+
+    @available(*, unavailable, renamed: "observe(_:)")
+    public func addNotificationBlock(_ block: @escaping (RealmCollectionChange<Self>) -> Void) -> NotificationToken {
+        fatalError()
+    }
 }