+++ RealmSwift/Tests/ObjectTests.swift
@@ -474,7 +474,7 @@ class ObjectTests: TestCase {
         try! realm.commitWrite()
 
         let exp = expectation(description: "")
-        let token = object.addNotificationBlock { change in
+        let token = object.observe { change in
             if case .deleted = change {
             } else {
                 XCTFail("expected .deleted, got \(change)")
@@ -487,7 +487,7 @@ class ObjectTests: TestCase {
         try! realm.commitWrite()
 
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
     }
 
     func expectChange<T: Equatable, U: Equatable>(_ name: String, _ old: T?, _ new: U?) -> ((ObjectChange) -> Void) {
@@ -513,13 +513,13 @@ class ObjectTests: TestCase {
         let object = realm.create(SwiftIntObject.self, value: [1])
         try! realm.commitWrite()
 
-        let token = object.addNotificationBlock(expectChange("intCol", Int?.none, 2))
+        let token = object.observe(expectChange("intCol", Int?.none, 2))
         try! realm.write {
             object.intCol = 2
         }
 
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
     }
 
     func testModifyObservedObjectRemotely() {
@@ -528,7 +528,7 @@ class ObjectTests: TestCase {
         let object = realm.create(SwiftIntObject.self, value: [1])
         try! realm.commitWrite()
 
-        let token = object.addNotificationBlock(expectChange("intCol", 1, 2))
+        let token = object.observe(expectChange("intCol", 1, 2))
         dispatchSyncNewThread {
             let realm = try! Realm()
             try! realm.write {
@@ -537,7 +537,7 @@ class ObjectTests: TestCase {
         }
 
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
     }
 
     func testListPropertyNotifications() {
@@ -546,7 +546,7 @@ class ObjectTests: TestCase {
         let object = realm.create(SwiftRecursiveObject.self, value: [[]])
         try! realm.commitWrite()
 
-        let token = object.addNotificationBlock(expectChange("objects", Int?.none, Int?.none))
+        let token = object.observe(expectChange("objects", Int?.none, Int?.none))
         dispatchSyncNewThread {
             let realm = try! Realm()
             try! realm.write {
@@ -556,7 +556,7 @@ class ObjectTests: TestCase {
         }
 
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
     }
 
     func testOptionalPropertyNotifications() {
@@ -566,7 +566,7 @@ class ObjectTests: TestCase {
             realm.add(object)
         }
 
-        var token = object.addNotificationBlock(expectChange("optIntCol", 1, 2))
+        var token = object.observe(expectChange("optIntCol", 1, 2))
         dispatchSyncNewThread {
             let realm = try! Realm()
             try! realm.write {
@@ -574,9 +574,9 @@ class ObjectTests: TestCase {
             }
         }
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
 
-        token = object.addNotificationBlock(expectChange("optIntCol", 2, Int?.none))
+        token = object.observe(expectChange("optIntCol", 2, Int?.none))
         dispatchSyncNewThread {
             let realm = try! Realm()
             try! realm.write {
@@ -584,9 +584,9 @@ class ObjectTests: TestCase {
             }
         }
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
 
-        token = object.addNotificationBlock(expectChange("optIntCol", Int?.none, 3))
+        token = object.observe(expectChange("optIntCol", Int?.none, 3))
         dispatchSyncNewThread {
             let realm = try! Realm()
             try! realm.write {
@@ -594,6 +594,6 @@ class ObjectTests: TestCase {
             }
         }
         waitForExpectations(timeout: 2)
-        token.stop()
+        token.invalidate()
     }
 }